optimized subscriptions page
